Thu, Aug 17, 2023, 7:24 PM -03:00

+++ readability/readability.c
@@ -32,9 +32,9 @@ int main(void)
     }
     int wor = words * 100;
     float L = letters/words * 100;
-    float S = sentences / wor;
+    int S = sentences / wor;
     int index = 0.0588 * L - 0.296 * S - 15.8;
-    printf("Grade %f\n", S);
+    printf("Grade %i\n", S);
 
 
     printf("%i\n", letters);
